+++ src/world/room.py
@@ -1,9 +1,13 @@
 import random
 
 import pygame
+from entity.excrement import Excrement
+from entity.grass import Grass
 from entity.living.bear import Bear
 from entity.living.chicken import Chicken
 from entity.living.livingEntity import LivingEntity
+from entity.stoneFloor import StoneFloor
+from entity.woodFloor import WoodFloor
 from lib.pyenvlib.environment import Environment
 from lib.graphik.src.graphik import Graphik
 
@@ -280,3 +284,63 @@ class Room(Environment):
 
     def setLivingEntities(self, livingEntities):
         self.livingEntities = livingEntities
+
+    def locationContainsEntityOfType(self, location, entityType):
+        for entityId in list(location.getEntities().keys()):
+            entity = location.getEntity(entityId)
+            if isinstance(entity, entityType):
+                return True
+        return False
+
+    def tickExcrement(self, tick, config):
+        # Living entities have a small chance to produce excrement
+        for entityId in self.livingEntities:
+            entity = self.livingEntities[entityId]
+            locationId = entity.getLocationID()
+            if str(locationId) == "-1":
+                continue
+            # 0.1% chance per tick
+            if random.randrange(1, 1001) > 1:
+                continue
+            try:
+                location = self.getGrid().getLocation(locationId)
+            except KeyError:
+                continue
+            excrement = Excrement(tick)
+            self.addEntityToLocation(excrement, location)
+
+        # Decay existing excrement into grass
+        entitiesToReplace = []
+        for locationId in self.getGrid().getLocations():
+            location = self.getGrid().getLocation(locationId)
+            expiredExcrement = []
+            for entityId in list(location.getEntities().keys()):
+                entity = location.getEntity(entityId)
+                if not isinstance(entity, Excrement):
+                    continue
+                if tick - entity.getTickCreated() < config.excrementDecayTicks:
+                    continue
+                expiredExcrement.append(entity)
+
+            if len(expiredExcrement) == 0:
+                continue
+
+            # Don't place grass on solid entities, existing grass, or floors
+            shouldPlaceGrass = True
+            if self.locationContainsSolidEntity(location):
+                shouldPlaceGrass = False
+            elif self.locationContainsEntityOfType(location, Grass):
+                shouldPlaceGrass = False
+            elif self.locationContainsEntityOfType(
+                location, StoneFloor
+            ) or self.locationContainsEntityOfType(location, WoodFloor):
+                shouldPlaceGrass = False
+
+            entitiesToReplace.append((location, expiredExcrement, shouldPlaceGrass))
+
+        for location, excrementList, shouldPlaceGrass in entitiesToReplace:
+            for excrement in excrementList:
+                location.removeEntity(excrement)
+            if shouldPlaceGrass:
+                grass = Grass()
+                self.addEntityToLocation(grass, location)